refactor: move to new routes directory
