+++ apps/server/src/trpc/v1/auth/signup.ts
@@ -1,6 +1,6 @@
 import { Auth } from "#lib/Auth.js";
 import Config from "#lib/Config.js";
-import { Utils } from "#lib/Auth.js";
+import { Utils } from "#lib/utils.js";
 import { createFieldError, publicProcedure, t } from "#trpc/lib.js";
 import { TRPCError } from "@trpc/server";
 import ms from "ms";
