+++ apps/web/src/app/files/[id]/auth/page.tsx
@@ -14,7 +14,7 @@ export function generateMetadata({ params }: { params: Record<string, string> })
 
 const Page: React.FC<PageProps<Params<"id">>> = async ({ params }) => {
 	const host = headers().get("host")!;
-	const status = await axios.get(`${getProtocol()}${host}/api/files/${params.id}`);
+	const status = await axios.get(`${getProtocol()}${host}/api/v1/files/${params.id}`);
 	if (typeof status.data === "object") redirect(`/files/${params.id}`);
 
 	return (
